TexFormulaParser.GetMatrixData → MatrixCommandParser.ReadMatrixCells (#100)
